refactor: moved supplementary functions for implementing controller functions into a separate helpers directory
