+++ backend/controllers/groupController.js
@@ -3,7 +3,10 @@
  * group data within a unit.
  *
  */
-const db_connection = require("../db_connection");
+const {promiseBasedQuery} = require("../helpers/commonHelpers");
+const {
+    groupFormationStrategies
+} = require("../helpers/groupControllerHelpers")
 
 const defaultGroupSize = 3;
 const defaultVariance = 1;
@@ -11,6 +14,8 @@ const defaultStrategy = "random";
 
 // get all groups from a unit
 const getAllGroups = async (req, res) => {
+    /* GET ALL GROUPS */
+    // '/:unitCode/:year/:period'
     res.status(200).send({wip: "test"});
 }
 
@@ -21,6 +26,15 @@ const getGroup = async (req, res) => {
 
 // create all the groups (based on csv)
 const createUnitGroups = async (req, res) => {
+    /* todo CURRENTLY RANDOMISE GROUPS */
+    // '/:unitCode/:year/:period'
+
+    /* GET ALL OF THE STUDENTS ASSOCIATED WITH THIS UNIT */
+    /* SORT THEM BY LAB NUMBER can do in the query */
+    /* RANDOMISE THE, STUDENTS WITHIN EACH LAB NUMBER */
+
+    /* SPLIT THE RANDOMISED LIST INTO GROUPS OF n AS SPECIFIED IN REQ */
+    /* FOR EACH GROUP,  */
     res.status(200).send({wip: "test"});
 }
 
@@ -31,6 +45,9 @@ const addGroup = async (req, res) => {
 
 // delete a specific group from a unit
 const deleteGroup = async (req, res) => {
+    /* DELETE THE STUDENT ALLOCATIONS BUT NOT STUDENTS */
+    /* DELETE THE GROUP ASSOCIATION WITH LAB/UNIT */
+    /* DELETE THE GROUP ITSELF */
     res.status(200).send({wip: "test"});
 }
 
@@ -40,44 +57,8 @@ const updateGroup = async (req, res) => {
 }
 
 const moveStudent = async (req, res) => {
+    /* REQUIRES OFFERING, STUDENT, OLD GROUP, NEW GROUP */
     res.status(200).send({wip: "test"});
-}
-
-/* SUPPLEMENTARY FUNCTIONS */
-/* ----------------------- */
-const promiseBasedQuery = (query, values) => {
-    /**
-     * wraps a mysql2 query around a promise so that we can use await with queries
-     */
-    return new Promise((resolve, reject) => {
-        db_connection.query(query, values, (error, results) => {
-            if (error) {
-                reject(error);
-            } else {
-                console.log(`${results.affectedRows} affected rows`)
-                resolve(results);
-            }
-        });
-    });
-}
-
-function shuffle(array) {
-    // Fisher-Yates shuffle algorithm from https://stackoverflow.com/questions/2450954/how-to-randomize-shuffle-a-javascript-array
-    let currentIndex = array.length, randomIndex;
-  
-    // While there remain elements to shuffle.
-    while (currentIndex != 0) {
-  
-      // Pick a remaining element.
-      randomIndex = Math.floor(Math.random() * currentIndex);
-      currentIndex--;
-  
-      // And swap it with the current element.
-      [array[currentIndex], array[randomIndex]] = [
-        array[randomIndex], array[currentIndex]];
-    }
-  
-    return array;
 }
 
 module.exports = {
